# project/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Logic Project - Member API
  version: "1.0.0"
  description: >
    회원가입/로그인/마이페이지/관리자 회원관리 기능의 API 명세서.
    실제 응답은 JSON이 아닌 Thymeleaf HTML 뷰(또는 302 리다이렉트)이며,
    Talend API Tester 등에서 URL/파라미터 테스트 용도로 사용하기 위해 작성했습니다.
servers:
  - url: http://localhost:8081
    description: 로컬 개발 서버

tags:
  - name: Home
  - name: Member
  - name: Admin

paths:
  /:
    get:
      tags: [Home]
      summary: 메인 페이지
      responses:
        "200":
          description: 메인 페이지 HTML

  /members/login:
    get:
      tags: [Member]
      summary: 로그인 화면
      responses:
        "200":
          description: 로그인 폼 HTML
    post:
      tags: [Member]
      summary: 로그인 처리
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [loginId, password]
              properties:
                loginId:
                  type: string
                  example: testuser
                password:
                  type: string
                  format: password
                  example: "Passw0rd!"
      responses:
        "302":
          description: 로그인 성공 시 '/' 로 리다이렉트 (세션에 loginMember 저장)
        "200":
          description: 로그인 실패(아이디 없음/비밀번호 불일치/비활성 계정) 시 에러 메시지와 함께 로그인 폼 재출력

  /members/logout:
    get:
      tags: [Member]
      summary: 로그아웃
      responses:
        "302":
          description: 세션 무효화 후 '/' 로 리다이렉트

  /members/terms:
    get:
      tags: [Member]
      summary: 이용약관 동의 화면
      responses:
        "200":
          description: 약관 HTML

  /members/signup:
    get:
      tags: [Member]
      summary: 회원가입 화면
      responses:
        "200":
          description: 회원가입 폼 HTML
    post:
      tags: [Member]
      summary: 회원가입 처리
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [loginId, password, name, email, termsAgreed, privacyAgreed]
              properties:
                loginId:
                  type: string
                  example: testuser
                password:
                  type: string
                  example: "Passw0rd!"
                name:
                  type: string
                  example: 홍길동
                email:
                  type: string
                  format: email
                  example: test@test.com
                phone:
                  type: string
                  example: "010-1234-5678"
                termsAgreed:
                  type: boolean
                  example: true
                privacyAgreed:
                  type: boolean
                  example: true
      responses:
        "302":
          description: 가입 성공 시 /members/login 으로 리다이렉트
        "200":
          description: 실패(아이디/이메일 중복, 약관 미동의 등) 시 에러 메시지와 함께 폼 재출력

  /members/mypage:
    get:
      tags: [Member]
      summary: 마이페이지 조회 (로그인 필요)
      responses:
        "200":
          description: 마이페이지 HTML
        "302":
          description: 미로그인 시 /members/login 으로 리다이렉트

  /members/edit:
    get:
      tags: [Member]
      summary: 회원정보 수정 화면 (로그인 필요)
      responses:
        "200":
          description: 수정 폼 HTML
        "302":
          description: 미로그인 시 /members/login 으로 리다이렉트
    post:
      tags: [Member]
      summary: 회원정보 수정 처리 (로그인 필요, 이름/이메일/전화번호만 변경)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [name, email, phone]
              properties:
                name:
                  type: string
                  example: 홍길동
                email:
                  type: string
                  format: email
                  example: new@test.com
                phone:
                  type: string
                  example: "010-9999-8888"
      responses:
        "302":
          description: 수정 성공 시 /members/mypage 로 리다이렉트

  /members/delete:
    get:
      tags: [Member]
      summary: 회원탈퇴 확인 화면 (로그인 필요)
      responses:
        "200":
          description: 탈퇴 확인 HTML
        "302":
          description: 미로그인 시 /members/login 으로 리다이렉트
    post:
      tags: [Member]
      summary: 회원탈퇴 처리 (로그인 필요, status를 KICK으로 변경하는 soft delete 후 세션 종료)
      responses:
        "302":
          description: 처리 후 '/' 로 리다이렉트

  /admin/members:
    get:
      tags: [Admin]
      summary: 전체 회원 목록 조회
      responses:
        "200":
          description: 회원 목록 HTML

  /admin/members/{id}:
    get:
      tags: [Admin]
      summary: 회원 상세 조회
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: 회원 상세 HTML

  /admin/members/{id}/activate:
    post:
      tags: [Admin]
      summary: 회원 활성화 (status = ACTIVE)
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "302":
          description: /admin/members 로 리다이렉트

  /admin/members/{id}/deactivate:
    post:
      tags: [Admin]
      summary: 회원 비활성화 (status = INACTIVE)
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "302":
          description: /admin/members 로 리다이렉트

  /admin/members/{id}/kick:
    post:
      tags: [Admin]
      summary: 회원 강제 탈퇴 (status = KICK)
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "302":
          description: /admin/members 로 리다이렉트

components:
  parameters:
    MemberId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 1
  schemas:
    Member:
      type: object
      properties:
        id:
          type: integer
          format: int64
        loginId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [USER, MANAGER, PART, ADMIN]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, KICK]
        createdAt:
          type: string
          format: date-time